split canvasUndoStack into drawing & textediting
