+++ ballontranslator/ui/image_edit.py
@@ -4,6 +4,10 @@ import cv2
 from qtpy.QtCore import QRectF, Qt, QPointF, QSize, QPoint, QDateTime
 from qtpy.QtWidgets import QStyleOptionGraphicsItem, QGraphicsPixmapItem, QWidget, QGraphicsPathItem, QGraphicsItem
 from qtpy.QtGui import QPen, QColor, QPainterPath, QCursor, QPainter, QPixmap, QImage, QBrush
+try:
+    from qtpy.QtWidgets import QUndoCommand
+except:
+    from qtpy.QtGui import QUndoCommand
 
 from .misc import DrawPanelConfig, pixmap2ndarray, ndarray2pixmap
 from utils.io_utils import imread, imwrite
@@ -159,4 +163,3 @@ class DrawingLayer(QGraphicsPixmapItem):
     def clearAllDrawings(self):
         self.qimg_dict.clear()
         self.drawing_items_info.clear()
-
