+++ ballontranslator/ui/page_search_widget.py
@@ -91,9 +91,7 @@ class HighlightMatched(QSyntaxHighlighter):
             if length > 0:
                 fmt.setBackground(CURRENT_TEXT_COLOR)
                 self.setFormat(intersect_start, length, fmt)
-        
         self.edit.highlighting = False
-        self.edit.updateUndoSteps()
 
 
 class SearchEditor(QPlainTextEdit):
@@ -690,140 +688,3 @@ class PageSearchWidget(Widget):
             else:
                 self.result_pos = 0
                 self.setCurrentEditor(edit)
-            
-
-class GlobalSearchWidget(Widget):
-    pass
-
-
-class ReplaceOneCommand(QUndoCommand):
-    def __init__(self, se: PageSearchWidget, parent=None):
-        super(ReplaceOneCommand, self).__init__(parent)
-        self.op_counter = 0
-        self.sw = se
-        self.reptxt = self.sw.replace_editor.toPlainText()
-        self.repl_len = len(self.reptxt)
-        
-        self.sel_start = self.sw.current_cursor.selectionStart()
-        self.oritxt = self.sw.current_cursor.selectedText()
-        self.ori_len = len(self.oritxt)
-        self.edit: Union[SourceTextEdit, TransTextEdit] = self.sw.current_edit
-        self.edit_is_src = type(self.edit) == SourceTextEdit
-        self.blkitem = self.sw.textblk_item_list[self.sw.current_edit.idx]
-
-        if self.sw.current_edit is not None and self.sw.isVisible():
-            move = self.sw.move_cursor(1)
-            if move == 0:
-                self.sw.result_pos = min(self.sw.counter_sum - 1, self.sw.result_pos + 1)
-            else:
-                self.sw.result_pos = 0
-
-        if not self.edit_is_src:
-            cursor = self.blkitem.textCursor()
-            cursor.setPosition(self.sel_start)
-            cursor.setPosition(self.sel_start+self.ori_len, QTextCursor.MoveMode.KeepAnchor)
-            cursor.beginEditBlock()
-            cursor.insertText(self.reptxt)
-            cursor.endEditBlock()
-
-        self.rep_cursor = self.edit.textCursor()
-        self.rep_cursor.setPosition(self.sel_start)
-        self.rep_cursor.setPosition(self.sel_start+self.ori_len, QTextCursor.MoveMode.KeepAnchor)
-        self.rep_cursor.insertText(self.reptxt)
-        self.edit.updateUndoSteps()
-
-    def redo(self):
-        if self.op_counter == 0:
-            self.op_counter += 1
-            return
-
-        if self.sw.current_edit is not None and self.sw.isVisible():
-            move = self.sw.move_cursor(1)
-            if move == 0:
-                self.sw.result_pos = min(self.sw.counter_sum - 1, self.sw.result_pos + 1)
-            else:
-                self.sw.result_pos = 0
-
-        if not self.edit_is_src:
-            self.blkitem.document().redo()
-        self.edit.redo()
-
-    def undo(self):
-        if not self.edit_is_src:
-            self.blkitem.document().undo()
-        self.edit.undo()
-
-        if self.sw.current_edit is not None and self.sw.isVisible():
-            move = self.sw.move_cursor(-1)
-            if move == 0:
-                self.sw.result_pos = max(self.sw.result_pos - 1, 0)
-            else:
-                self.sw.result_pos = self.sw.counter_sum - 1
-            self.sw.updateCounterText()
-
-        self.edit.user_edited.emit()
-
-class ReplaceAllCommand(QUndoCommand):
-
-    def __init__(self, search_widget: PageSearchWidget) -> None:
-        super().__init__()
-        self.op_counter = 0
-        self.sw = search_widget
-
-        self.rstedit_list: List[SourceTextEdit] = []
-        self.blkitem_list: List[TextBlkItem] = []
-        curpos_list = []
-        for edit, highlighter in zip(self.sw.search_rstedit_list, self.sw.highlighter_list):
-            self.rstedit_list.append(edit)
-            curpos_list.append(list(highlighter.matched_map.keys()))
-
-        text = self.sw.search_editor.toPlainText()
-        len_text = len(text)
-        replace = self.sw.replace_editor.toPlainText()
-        len_delta = len(replace) - len_text
-        for edit, curpos_lst in zip(self.rstedit_list, curpos_list):
-            redo_blk = type(edit) == TransTextEdit
-            if redo_blk:
-                blkitem = self.sw.textblk_item_list[edit.idx]
-                self.blkitem_list.append(blkitem)
-                sel_list = []
-
-            cursor = edit.textCursor()
-            cursor.clearSelection()
-            cursor.setPosition(0)
-            cursor.beginEditBlock()
-            for ii, sel_start in enumerate(curpos_lst):
-                sel_start += len_delta * ii - len_text
-                cursor.setPosition(sel_start)
-                sel_end = sel_start+len_text
-                cursor.setPosition(sel_end, QTextCursor.MoveMode.KeepAnchor)
-                cursor.insertText(replace)
-                if redo_blk:
-                    sel_list.append([sel_start, sel_end])
-            cursor.endEditBlock()
-            edit.updateUndoSteps()
-
-            if redo_blk:
-                cursor = blkitem.textCursor()
-                cursor.beginEditBlock()
-                for sel in sel_list:
-                    cursor.setPosition(sel[0])
-                    cursor.setPosition(sel[1], QTextCursor.MoveMode.KeepAnchor)
-                    cursor.insertText(replace)
-                cursor.endEditBlock()
-
-    def redo(self):
-        if self.op_counter == 0:
-            self.op_counter += 1
-            return
-
-        for edit in self.rstedit_list:
-            edit.redo()
-        for blkitem in self.blkitem_list:
-            blkitem.document().redo()
-
-    def undo(self):
-        for edit in self.rstedit_list:
-            edit.undo()
-        for blkitem in self.blkitem_list:
-            blkitem.document().undo()
