+++ ballontranslator/ui/scenetext_manager.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from qtpy.QtWidgets import QApplication
 from qtpy.QtCore import QObject, QRectF, Qt, Signal
-from qtpy.QtGui import QTextCursor, QFontMetrics, QFont, QTextCharFormat, QInputMethodEvent
+from qtpy.QtGui import QTextCursor, QFontMetrics, QFont, QTextCharFormat
 try:
     from qtpy.QtWidgets import QUndoCommand
 except:
@@ -12,165 +12,13 @@ except:
 
 from .textitem import TextBlkItem, TextBlock, xywh2xyxypoly
 from .canvas import Canvas
-from .textedit_area import TextPanel, TransTextEdit, SourceTextEdit, TransPairWidget, TextEditCommand
+from .textedit_area import TextPanel, TransTextEdit, SourceTextEdit, TransPairWidget
 from .fontformatpanel import set_textblk_fontsize
 from .misc import FontFormat, ProgramConfig, pt2px
-from .texteditshapecontrol import TextBlkShapeControl
-
+from .textedit_commands import propagate_user_edit, TextEditCommand, ReshapeItemCommand, MoveBlkItemsCommand, AutoLayoutCommand, ApplyFontformatCommand, ApplyEffectCommand, RotateItemCommand, TextItemEditCommand, TextEditCommand, PageReplaceOneCommand, PageReplaceAllCommand
 from utils.imgproc_utils import extract_ballon_region
 from utils.text_processing import seg_text, is_cjk
 from utils.text_layout import layout_text
-
-class MoveBlkItemsCommand(QUndoCommand):
-    def __init__(self, items: List[TextBlkItem], shape_ctrl: TextBlkShapeControl):
-        super(MoveBlkItemsCommand, self).__init__()
-        self.items = items
-        self.old_pos_lst = []
-        self.new_pos_lst = []
-        self.shape_ctrl = shape_ctrl
-        for item in items:
-            self.old_pos_lst.append(item.oldPos)
-            self.new_pos_lst.append(item.pos())
-            item.oldPos = item.pos()
-
-    def redo(self):
-        for item, new_pos in zip(self.items, self.new_pos_lst):
-            item.setPos(new_pos)
-            if self.shape_ctrl.blk_item == item and self.shape_ctrl.pos() != new_pos:
-                self.shape_ctrl.setPos(new_pos)
-
-    def undo(self):
-        for item, old_pos in zip(self.items, self.old_pos_lst):
-            item.setPos(old_pos)
-            if self.shape_ctrl.blk_item == item and self.shape_ctrl.pos() != old_pos:
-                self.shape_ctrl.setPos(old_pos)
-
-    def mergeWith(self, command: QUndoCommand):
-        if command.old_pos_lst == self.old_pos_lst:
-            return True
-        return False
-
-
-class ApplyFontformatCommand(QUndoCommand):
-    def __init__(self, items: List[TextBlkItem], fontformat: FontFormat):
-        super(ApplyFontformatCommand, self).__init__()
-        self.items = items
-        self.old_html_lst = []
-        self.old_rect_lst = []
-        self.old_fmt_lst = []
-        self.new_fmt = fontformat
-        for item in items:
-            self.old_html_lst.append(item.toHtml())
-            self.old_fmt_lst.append(item.get_fontformat())
-            self.old_rect_lst.append(item.absBoundingRect())
-
-    def redo(self):
-        for item in self.items:
-            item.set_fontformat(self.new_fmt, set_char_format=True)
-
-    def undo(self):
-        for rect, item, html, fmt in zip(self.old_rect_lst, self.items, self.old_html_lst, self.old_fmt_lst):
-            item.setHtml(html)
-            item.set_fontformat(fmt)
-            item.setRect(rect)
-
-    def mergeWith(self, command: QUndoCommand):
-        if command.new_fmt == self.new_fmt:
-            return True
-        return False
-
-
-class ApplyEffectCommand(QUndoCommand):
-    def __init__(self, items: List[TextBlkItem], fontformat: FontFormat):
-        super(ApplyEffectCommand, self).__init__()
-        self.items = items
-        self.old_fmt_lst: List[FontFormat] = []
-        self.new_fmt = fontformat
-        for item in items:
-            self.old_fmt_lst.append(item.get_fontformat())
-
-    def redo(self):
-        for item in self.items:
-            item.update_effect(self.new_fmt)
-            item.update()
-
-    def undo(self):
-        for item, fmt in zip(self.items, self.old_fmt_lst):
-            item.update_effect(fmt)
-            item.update()
-
-class ReshapeItemCommand(QUndoCommand):
-    def __init__(self, item: TextBlkItem, parent=None):
-        super(ReshapeItemCommand, self).__init__(parent)
-        self.item = item
-        self.oldRect = item.oldRect
-        self.newRect = item.absBoundingRect()
-
-    def redo(self):
-        self.item.setRect(self.newRect)
-
-    def undo(self):
-        self.item.setRect(self.oldRect)
-
-    def mergeWith(self, command: QUndoCommand):
-        item = command.item
-        if self.item != item:
-            return False
-        self.newRect = item.rect()
-        return True
-
-
-class RotateItemCommand(QUndoCommand):
-    def __init__(self, item: TextBlkItem, new_angle: float, shape_ctrl: TextBlkShapeControl):
-        super(RotateItemCommand, self).__init__()
-        self.item = item
-        self.old_angle = item.rotation()
-        self.new_angle = new_angle
-        self.shape_ctrl = shape_ctrl
-
-    def redo(self):
-        self.item.setRotation(self.new_angle)
-        self.item.blk.angle = self.new_angle
-        if self.shape_ctrl.blk_item == self.item and self.shape_ctrl.rotation() != self.new_angle:
-            self.shape_ctrl.setRotation(self.new_angle)
-
-    def undo(self):
-        self.item.setRotation(self.old_angle)
-        self.item.blk.angle = self.old_angle
-        if self.shape_ctrl.blk_item == self.item and self.shape_ctrl.rotation() != self.old_angle:
-            self.shape_ctrl.setRotation(self.old_angle)
-
-    def mergeWith(self, command: QUndoCommand):
-        item = command.item
-        if self.item != item:
-            return False
-        self.new_angle = item.angle
-        return True
-
-
-class OrientationItemCommand(QUndoCommand):
-    def __init__(self, item: TextBlkItem, ctrl):
-        super(OrientationItemCommand, self).__init__()
-        self.item = item
-        self.ctrl: SceneTextManager = ctrl
-        self.oldVertical = item.is_vertical
-        self.newVertical = self.ctrl.fontformat.vertical
-
-    def redo(self):
-        self.item.setVertical(self.newVertical)
-        self.ctrl.formatpanel.verticalChecker.setChecked(self.newVertical)
-
-    def undo(self):
-        self.item.setVertical(self.oldVertical)
-        self.ctrl.formatpanel.verticalChecker.setChecked(self.oldVertical)
-
-    def mergeWith(self, command: QUndoCommand):
-        item = command.item
-        if self.item != item:
-            return False
-        self.newVertical = command.newVertical
-        self.oldVertical = command.oldVertical
-        return True
 
 
 class CreateItemCommand(QUndoCommand):
@@ -279,7 +127,6 @@ class DeleteBlkItemsCommand(QUndoCommand):
                 else:
                     self.sw.setCurrentEditor(None)
 
-
     def undo(self):
         self.ctrl.recoverTextblkItemList(self.blk_list, self.pwidget_list)
         if self.sw_changed:
@@ -288,71 +135,6 @@ class DeleteBlkItemsCommand(QUndoCommand):
             self.sw.search_counter_list += self.search_counter_list
             self.sw.highlighter_list += self.highlighter_list
             self.sw.updateCounterText()
-
-
-class AutoLayoutCommand(QUndoCommand):
-    def __init__(self, items: List[TextBlkItem], old_rect_lst: List, old_html_lst: List, trans_widget_lst: List[TransTextEdit]):
-        super(AutoLayoutCommand, self).__init__()
-        self.items = items
-        self.old_html_lst = old_html_lst
-        self.old_rect_lst = old_rect_lst
-        self.trans_widget_lst = trans_widget_lst
-        self.new_rect_lst = []
-        self.new_html_lst = []
-        for item in items:
-            self.new_html_lst.append(item.toHtml())
-            self.new_rect_lst.append(item.absBoundingRect())
-        self.counter = 0
-
-    def redo(self):
-        self.counter += 1
-        if self.counter <= 1:
-            return
-        for item, trans_widget, html, rect  in zip(self.items, self.trans_widget_lst, self.new_html_lst, self.new_rect_lst):
-            trans_widget.setPlainText(item.toPlainText())
-            item.setPlainText('')
-            item.setRect(rect, repaint=False)
-            item.setHtml(html)
-            if item.letter_spacing != 1:
-                item.setLetterSpacing(item.letter_spacing, force=True)
-            
-
-    def undo(self):
-        for item, trans_widget, html, rect  in zip(self.items, self.trans_widget_lst, self.old_html_lst, self.old_rect_lst):
-            trans_widget.setPlainText(item.toPlainText())
-            item.setPlainText('')
-            item.setRect(rect, repaint=False)
-            item.setHtml(html)
-            if item.letter_spacing != 1:
-                item.setLetterSpacing(item.letter_spacing, force=True)
-
-
-class TextItemEditCommand(QUndoCommand):
-    def __init__(self, blkitem: TextBlkItem, trans_edit: TransTextEdit, num_steps: int):
-        super(TextItemEditCommand, self).__init__()
-        self.op_counter = 0
-        self.edit = trans_edit
-        self.blkitem = blkitem
-        self.num_steps = min(num_steps, 2)
-        if blkitem.input_method_from == -1:
-            self.num_steps = 1
-        else:
-            blkitem.input_method_from = -1
-
-    def redo(self):
-        if self.op_counter == 0:
-            self.op_counter += 1
-            return
-        for _ in range(self.num_steps):
-            self.blkitem.redo()
-        if self.edit is not None:
-            self.edit.redo()
-
-    def undo(self):
-        for _ in range(self.num_steps):
-            self.blkitem.undo()
-        if self.edit is not None:
-            self.edit.undo()
 
 
 class SceneTextManager(QObject):
@@ -370,7 +152,6 @@ class SceneTextManager(QObject):
         self.canvas.delete_textblks.connect(self.onDeleteBlkItems)
         self.canvas.format_textblks.connect(self.onFormatTextblks)
         self.canvas.layout_textblks.connect(self.onAutoLayoutTextblks)
-        self.canvasUndoStack = self.canvas.undoStack
         self.txtblkShapeControl = canvas.txtblkShapeControl
         self.textpanel = textpanel
 
@@ -447,7 +228,6 @@ class SceneTextManager(QObject):
         self.pairwidget_list.append(pair_widget)
         self.textEditList.addPairWidget(pair_widget)
         pair_widget.e_source.setPlainText(blk_item.blk.get_text())
-        pair_widget.e_source.user_edited.connect(self.on_srcwidget_edited)
         pair_widget.e_source.ensure_scene_visible.connect(self.on_ensure_textitem_svisible)
         pair_widget.e_source.push_undo_stack.connect(self.on_push_edit_stack)
         pair_widget.e_source.redo_signal.connect(self.on_textedit_redo)
@@ -455,7 +235,7 @@ class SceneTextManager(QObject):
 
         pair_widget.e_trans.setPlainText(blk_item.toPlainText())
         pair_widget.e_trans.focus_in.connect(self.onTransWidgetHoverEnter)
-        pair_widget.e_trans.user_edited_verbose.connect(self.onTransWidgetUserEditedVerbose)
+        pair_widget.e_trans.propagate_user_edited.connect(self.on_propagate_transwidget_edit)
         pair_widget.e_trans.ensure_scene_visible.connect(self.on_ensure_textitem_svisible)
         pair_widget.e_trans.push_undo_stack.connect(self.on_push_edit_stack)
         pair_widget.e_trans.redo_signal.connect(self.on_textedit_redo)
@@ -478,7 +258,7 @@ class SceneTextManager(QObject):
         textblk_item.push_undo_stack.connect(self.on_push_textitem_undostack)
         textblk_item.undo_signal.connect(self.on_textedit_undo)
         textblk_item.redo_signal.connect(self.on_textedit_redo)
-        textblk_item.user_edited_verbose.connect(self.onTextItemUserEditedVerbose)
+        textblk_item.propagate_user_edited.connect(self.on_propagate_textitem_edit)
         textblk_item.doc_size_changed.connect(self.onTextBlkItemSizeChanged)
         textblk_item.pasted.connect(self.onBlkitemPaste)
         return textblk_item
@@ -516,7 +296,6 @@ class SceneTextManager(QObject):
             if self.txtblkShapeControl.blk_item is not None and blkitem.isSelected():
                 blkitem.setSelected(False)
         
-
     def onTextBlkItemSizeChanged(self, idx: int):
         blk_item = self.textblk_item_list[idx]
         if not self.txtblkShapeControl.reshaping:
@@ -587,22 +366,22 @@ class SceneTextManager(QObject):
     def onTextBlkItemMoved(self):
         selected_blks = self.get_selected_blkitems()
         if len(selected_blks) > 0:
-            self.canvasUndoStack.push(MoveBlkItemsCommand(selected_blks, self.txtblkShapeControl))
+            self.canvas.push_undo_command(MoveBlkItemsCommand(selected_blks, self.txtblkShapeControl))
         
     def onTextBlkItemReshaped(self, item: TextBlkItem):
-        self.canvasUndoStack.push(ReshapeItemCommand(item))
+        self.canvas.push_undo_command(ReshapeItemCommand(item))
 
     def onTextBlkItemRotated(self, new_angle: float):
         blk_item = self.txtblkShapeControl.blk_item
         if blk_item:
-            self.canvasUndoStack.push(RotateItemCommand(blk_item, new_angle, self.txtblkShapeControl))
+            self.canvas.push_undo_command(RotateItemCommand(blk_item, new_angle, self.txtblkShapeControl))
 
     def onDeleteBlkItems(self):
         selected_blks = self.get_selected_blkitems()
         if len(selected_blks) == 0 and self.txtblkShapeControl.blk_item is not None:
             selected_blks.append(self.txtblkShapeControl.blk_item)
         if len(selected_blks) > 0:
-            self.canvasUndoStack.push(DeleteBlkItemsCommand(selected_blks, self))
+            self.canvas.push_undo_command(DeleteBlkItemsCommand(selected_blks, self))
 
     def onFormatTextblks(self):
         self.apply_fontformat(self.formatpanel.global_format)
@@ -618,7 +397,7 @@ class SceneTextManager(QObject):
                 trans_widget_lst.append(self.pairwidget_list[blkitem.idx].e_trans)
                 self.layout_textblk(blkitem)
 
-            self.canvasUndoStack.push(AutoLayoutCommand(selected_blks, old_rect_lst, old_html_lst, trans_widget_lst))
+            self.canvas.push_undo_command(AutoLayoutCommand(selected_blks, old_rect_lst, old_html_lst, trans_widget_lst))
 
     def layout_textblk(self, blkitem: TextBlkItem, text: str = None, mask: np.ndarray = None, bounding_rect: List = None, region_rect: List = None):
         
@@ -817,10 +596,10 @@ class SceneTextManager(QObject):
             block.lines = xywh2xyxypoly(np.array([xywh])).reshape(-1, 4, 2).tolist()
             blk_item = TextBlkItem(block, len(self.textblk_item_list), set_format=False, show_rect=True)
             blk_item.set_fontformat(self.formatpanel.global_format)
-            self.canvasUndoStack.push(CreateItemCommand(blk_item, self))
+            self.canvas.push_undo_command(CreateItemCommand(blk_item, self))
 
     def onRotateTextBlkItem(self, item: TextBlock):
-        self.canvasUndoStack.push(RotateItemCommand(item))
+        self.canvas.push_undo_command(RotateItemCommand(item))
     
     def onTransWidgetHoverEnter(self, idx: int):
         if self.is_editting():
@@ -830,77 +609,46 @@ class SceneTextManager(QObject):
         self.txtblkShapeControl.setBlkItem(blk_item)
 
     def on_textedit_redo(self):
-        self.canvasUndoStack.redo()
+        self.canvas.redo_textedit()
 
     def on_textedit_undo(self):
-        self.canvasUndoStack.undo()
+        self.canvas.undo_textedit()
 
     def on_push_textitem_undostack(self, num_steps: int, is_formatting: bool):
         blkitem: TextBlkItem = self.sender()
         e_trans = self.pairwidget_list[blkitem.idx].e_trans if not is_formatting else None
-        self.canvasUndoStack.push(TextItemEditCommand(blkitem, e_trans, num_steps))
+        self.canvas.push_undo_command(TextItemEditCommand(blkitem, e_trans, num_steps))
 
     def on_push_edit_stack(self, num_steps: int):
         edit: Union[TransTextEdit, SourceTextEdit] = self.sender()
         blkitem = self.textblk_item_list[edit.idx] if type(edit) == TransTextEdit else None
-        self.canvasUndoStack.push(TextEditCommand(edit, num_steps, blkitem))
+        self.canvas.push_undo_command(TextEditCommand(edit, num_steps, blkitem))
 
-    def onTextItemUserEditedVerbose(self, pos: int, added_text: str, input_method_used: bool):
+    def on_propagate_textitem_edit(self, pos: int, added_text: str, input_method_used: bool):
         blk_item: TextBlkItem = self.sender()
-        idx = blk_item.idx
-        edit = self.pairwidget_list[idx].e_trans
+        edit = self.pairwidget_list[blk_item.idx].e_trans
+        propagate_user_edit(blk_item, edit, pos, added_text, input_method_used)
 
-        ori_count = edit.document().characterCount()
-        new_count = blk_item.document().characterCount()
-        removed = ori_count + len(added_text) - new_count
-
-        cursor = edit.textCursor()
-        cursor.setPosition(pos)
-        if removed > 0:
-            cursor.setPosition(pos + removed, QTextCursor.MoveMode.KeepAnchor)
-        if input_method_used:
-            cursor.beginEditBlock()
-        cursor.insertText((added_text))
-        if input_method_used:
-            cursor.endEditBlock()
-        self.canvas.setProjSaveState(True)
-
-    def onTransWidgetUserEditedVerbose(self, pos: int, added_text: str, input_method_used: bool):
+    def on_propagate_transwidget_edit(self, pos: int, added_text: str, input_method_used: bool):
         edit: TransTextEdit = self.sender()
-        idx = edit.idx
-        blk_item = self.textblk_item_list[idx]
+        blk_item = self.textblk_item_list[edit.idx]
         if blk_item.isEditing():
             blk_item.setTextInteractionFlags(Qt.TextInteractionFlag.NoTextInteraction)
-
-        ori_count = blk_item.document().characterCount()
-        new_count = edit.document().characterCount()
-        removed = ori_count + len(added_text) - new_count
-
-        cursor = blk_item.textCursor()
-        cursor.setPosition(pos)
-        if removed > 0:
-            cursor.setPosition(pos + removed, QTextCursor.MoveMode.KeepAnchor)
-        if input_method_used:
-            cursor.beginEditBlock()
-        cursor.insertText((added_text))
-        if input_method_used:
-            cursor.endEditBlock()
-        self.canvas.setProjSaveState(True)
+        propagate_user_edit(edit, blk_item, pos, added_text, input_method_used)
 
     def onGlobalFormatChanged(self):
-        # if not isinstance(self.app.focusWidget(), PaintQSlider):
         self.apply_fontformat(self.formatpanel.global_format)
 
     def apply_fontformat(self, fontformat: FontFormat):
         selected_blks = self.get_selected_blkitems()
         if len(selected_blks) > 0:
-            self.canvasUndoStack.push(ApplyFontformatCommand(selected_blks, fontformat))
+            self.canvas.push_undo_command(ApplyFontformatCommand(selected_blks, fontformat))
 
     def on_apply_effect(self):
         format = self.formatpanel.active_format
         selected_blks = self.get_selected_blkitems()
         if len(selected_blks) > 0:
-            self.canvasUndoStack.push(ApplyEffectCommand(selected_blks, format))
+            self.canvas.push_undo_command(ApplyEffectCommand(selected_blks, format))
 
     def get_selected_blkitems(self) -> List[TextBlkItem]:
         selections = self.canvas.selectedItems()
@@ -909,9 +657,6 @@ class SceneTextManager(QObject):
             if isinstance(selection, TextBlkItem):
                 selected_blks.append(selection)
         return selected_blks
-
-    def on_srcwidget_edited(self):
-        self.canvas.setProjSaveState(True)
 
     def updateTextBlkItemIdx(self):
         for ii, blk_item in enumerate(self.textblk_item_list):
@@ -939,6 +684,7 @@ class SceneTextManager(QObject):
         for blk_item, transwidget in zip(self.textblk_item_list, self.pairwidget_list):
             transwidget.e_trans.setPlainText(blk_item.blk.translation)
             blk_item.setPlainText(blk_item.blk.translation)
+        self.canvas.clear_text_stack()
 
     def showTextblkItemRect(self, draw_rect: bool):
         for blk_item in self.textblk_item_list:
@@ -957,6 +703,12 @@ class SceneTextManager(QObject):
         self.canvas.gv.ensureVisible(self.textblk_item_list[edit.idx])
         self.txtblkShapeControl.setBlkItem(self.textblk_item_list[edit.idx])
 
+    def on_page_replace_one(self):
+        self.canvas.push_undo_command(PageReplaceOneCommand(self.canvas.search_widget))
+
+    def on_page_replace_all(self):
+        self.canvas.push_undo_command(PageReplaceAllCommand(self.canvas.search_widget))
+
 def get_text_size(fm: QFontMetrics, text: str) -> Tuple[int, int]:
     brt = fm.tightBoundingRect(text)
     br = fm.boundingRect(text)
@@ -964,3 +716,4 @@ def get_text_size(fm: QFontMetrics, text: str) -> Tuple[int, int]:
     
 def get_words_length_list(fm: QFontMetrics, words: List[str]) -> List[int]:
     return [fm.tightBoundingRect(word).width() for word in words]
+
